@@ -13,35 +13,35 @@
         }
     ],
     "require": {
-        "php": "^7.3",
+        "php": "^7.4",
         "ext-PDO": "*",
         "ext-bz2": "*",
         "ext-json": "*",
         "ext-zip": "*",
-        "doctrine/annotations": "^1.10.2",
-        "egulias/email-validator": "^2.1.17",
-        "fzaninotto/faker": "^1.9.1",
-        "google/apiclient": "^2.4.1",
-        "google/cloud": "^0.123.0",
-        "google/protobuf": "^v3.11.4",
+        "doctrine/annotations": "^1.12.1",
+        "egulias/email-validator": "^2.1.25",
+        "fzaninotto/faker": "^v1.9.2",
+        "google/apiclient": "^v2.9.1",
+        "google/cloud": "^v0.150.0",
+        "google/protobuf": "^v3.15.2",
         "google/recaptcha": "^1.2.4",
         "grpc/grpc": "^v1.27.0",
         "jolicode/slack-php-api": "^2.5",
-        "kreait/firebase-php": "^4.41.0",
-        "lcobucci/jwt": "^3.3.1",
-        "nesbot/carbon": "^2.32.2",
-        "nyholm/psr7": "^1.3",
-        "php-di/php-di": "^6.1.0",
-        "phpmailer/phpmailer": "^6.1.5",
+        "kreait/firebase-php": "^5.8.1",
+        "lcobucci/jwt": "^4.1.4",
+        "nesbot/carbon": "^2.45.1",
+        "nyholm/psr7": "^1.4.0",
+        "php-di/php-di": "^6.3.0",
+        "phpmailer/phpmailer": "^v6.3.0",
         "ramsey/uuid": "^3.9.3",
-        "robmorgan/phinx": "^0.11.6",
-        "sendgrid/sendgrid": "^7.5.0",
+        "robmorgan/phinx": "^0.12.5",
+        "sendgrid/sendgrid": "^7.9.2",
         "slim/php-view": "^2.2.1",
-        "slim/psr7": "^1.1.0",
-        "slim/slim": "^4.5.0",
-        "symfony/http-client": "^5.1",
-        "symfony/validator": "^5.0.8",
-        "zircote/swagger-php": "^3.0.4"
+        "slim/psr7": "^1.3.0",
+        "slim/slim": "^4.7.1",
+        "symfony/http-client": "^v5.2.3",
+        "symfony/validator": "^v5.1.5",
+        "zircote/swagger-php": "^3.1.0"
     },
     "autoload": {
       "psr-4": {"RedCrossQuest\\": "src/"}
@@ -55,7 +55,7 @@
             "*": "dist"
         },
         "platform": {
-            "php": "7.3"
+            "php": "7.4"
         }
     }
 }
