@@ -1,14 +1,10 @@
 <?php
-
-
 namespace RedCrossQuest\routes\routesActions\authentication;
 
-
+use DateTimeImmutable;
 use DI\Annotation\Inject;
 use Google\ApiCore\ApiException;
-use Lcobucci\JWT\Builder;
-use Lcobucci\JWT\Signer\Hmac\Sha256;
-use Lcobucci\JWT\Signer\Key;
+use Lcobucci\JWT\Configuration;
 use Lcobucci\JWT\Token;
 use Psr\Log\LoggerInterface;
 use RedCrossQuest\Entity\QueteurEntity;
@@ -16,7 +12,6 @@ use RedCrossQuest\Entity\UniteLocaleEntity;
 use RedCrossQuest\Entity\UserEntity;
 use RedCrossQuest\routes\routesActions\Action;
 use RedCrossQuest\Service\ClientInputValidator;
-use RedCrossQuest\Service\SecretManagerService;
 
 
 abstract class AuthenticateAbstractAction extends Action
@@ -25,22 +20,22 @@ abstract class AuthenticateAbstractAction extends Action
    * @Inject("settings")
    * @var array settings
    */
-  protected $settings;
-  /** @var SecretManagerService    $secretManagerService*/
-  private $secretManagerService;
+  protected array $settings;
+  /** @var Configuration           $JWTConfiguration*/
+  private Configuration           $JWTConfiguration;
 
 
   /**
    * @param LoggerInterface $logger
    * @param ClientInputValidator $clientInputValidator
-   * @param SecretManagerService $secretManagerService
+   * @param Configuration           $JWTConfiguration
    */
   public function __construct(LoggerInterface         $logger,
                               ClientInputValidator    $clientInputValidator,
-                              SecretManagerService    $secretManagerService)
+                              Configuration           $JWTConfiguration)
   {
     parent::__construct($logger, $clientInputValidator);
-    $this->secretManagerService = $secretManagerService;
+    $this->JWTConfiguration = $JWTConfiguration;
   }
 
 
@@ -55,21 +50,20 @@ abstract class AuthenticateAbstractAction extends Action
    */
   protected function getToken(QueteurEntity $queteur, UniteLocaleEntity $ul, UserEntity $user) : Token
   {
-    $jwtSecret      = $this->secretManagerService->getSecret(SecretManagerService::$JWT_SECRET);
+
     $issuer         = $this->settings['jwt'        ]['issuer'        ];
     $audience       = $this->settings['jwt'        ]['audience'      ];
     $deploymentType = $this->settings['appSettings']['deploymentType'];
     $sessionLength  = $this->settings['appSettings']['sessionLength' ];
 
-    $signer   = new Sha256();
-    $issuedAt =     time  ();
+    $now    = new DateTimeImmutable();
 
-    return (new Builder())
+    return $this->JWTConfiguration->builder()
       ->issuedBy          ($issuer  ) // Configures the issuer (iss claim)
       ->permittedFor      ($audience)
-      ->issuedAt          ($issuedAt)
-      ->canOnlyBeUsedAfter($issuedAt)
-      ->expiresAt         (time() + $sessionLength * 3600)
+      ->issuedAt          ($now)
+      ->canOnlyBeUsedAfter($now->modify('+1 minute'))
+      ->expiresAt         ($now->modify("+$sessionLength hour"))
       ->withClaim         ('username' , $queteur->nivol)
       ->withClaim         ('id'       , $user->id      )
       ->withClaim         ('ulId'     , $queteur->ul_id)
@@ -78,6 +72,6 @@ abstract class AuthenticateAbstractAction extends Action
       ->withClaim         ('queteurId', $queteur->id   )
       ->withClaim         ('roleId'   , $user->role    )
       ->withClaim         ('d'        , $deploymentType)
-      ->getToken          ($signer          , new Key($jwtSecret));
+      ->getToken          ($this->JWTConfiguration->signer() , $this->JWTConfiguration->signingKey());
   }
 }
